[FIX] test_invader_payment: Direct call of the method on the abract component
invader_payment_* addons provides abstract services. Therefore, the service's methods are not publicly exposed as a REST api since no controller is registered to serve these methods. To test these methodes you must call them directly. Indeed, the generic dispatch method can't be used to access these methods in the current context since it checks if a route is associated to ensure that the method is public. (Change required by the new implemantation of base_rest)